E-shop main functionallieties implemented and working properly, updated page design

+++ app/Http/Controllers/AdminMediaConrtoller.php
@@ -31,7 +31,9 @@ class AdminMediaConrtoller extends Controller
 
     public function destroy($id){
       $file = Photo::findOrFail($id);
-      unlink(public_path() . $file->path);
+      if(file_exists(public_path().$file->path)){
+        unlink(public_path() . $file->path);
+      }
       $file->delete();
       Session::flash('message', 'File ' .$file->path. ' has been deleted');
       return redirect('/admin/media');
